Correction of all remarks
